requirements.txt
pinecone
python-dotenv
fastapi
uvicorn[standard]
pydantic
pytest
httpx

fastapi
uvicorn
pymupdf
pandas
pytesseract
moviepy
openai
sentence-transformers
faiss-cpu
python-multipart
Pillow
